backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
langchain>=0.0.267
langchain-google-genai>=0.0.5
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.3
pandas>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=1.10.7
aiofiles>=23.1.0
pillow>=9.5.0
transformers>=4.30.2